Move netcore LocalDBAPI.Common into LocalDBAPI.Windows since it is only loaded on Windows
